Set MSB in salt for tunnel password instead of erroring out
The code in NewTunnelPassword() currently checks for the MSB in the salt to
be set and errors out otherwise. This is problematic as the users of this
code are generated code, and they all simply provide a random 2-octet value
that may or may not follow that requirement. This means that the functions
will fail in half of all calls.

Instead of erroring out in such a case, just set the bit to obey to the rfc.