bugfixed getstring method for tagged attributes
The library had a bug where the `Get()` function will return the correct
tag, but the `GetString()` function will not. This is because the
returned tag is being ignored when passing the result of `LookupString()`
to `GetString()`.
This commit fixes the problem, so that the response is not ignored
anymore and you're able to access the tag value.